+++ Quicksilver/Code-External/NDClasses/NSString+NDUtilities.m
@@ -112,7 +112,7 @@
 /*
 	-indexOfCharacter:range:
  */
-- (unsigned int)indexOfCharacter:(unichar)aCharacter range:(NSRange)aRange
+- (NSUInteger)indexOfCharacter:(unichar)aCharacter range:(NSRange)aRange
 {
 	NSUInteger theIndex, theCount = [self length], theFoundIndex = NSNotFound;
 
@@ -131,7 +131,7 @@
 /*
 	-indexOfCharacter:
  */
-- (unsigned int)indexOfCharacter:(unichar)aCharacter
+- (NSUInteger)indexOfCharacter:(unichar)aCharacter
 {
 	return [self indexOfCharacter:aCharacter range:NSMakeRange( 0, [self length] )];
 }
@@ -203,7 +203,7 @@
 /*
 	-indexOfMatchingStringInList:defaultValue:
  */
-- (unsigned int)indexOfMatchingStringInList:(NSString **)anArray defaultValue:(unsigned int)aDefaultValue
+- (NSUInteger)indexOfMatchingStringInList:(NSString **)anArray defaultValue:(NSUInteger)aDefaultValue
 {
 	unsigned int		theIndex = 0,
 						theFoundIndex = UINT_MAX;
@@ -236,7 +236,7 @@
 - (NSRange)rangeOfStringEnclosedIn:(NSString *)aStartString and:(NSString *)anEndString includeEncloseString:(BOOL)anIncludeEnclose mode:(int)aMode
 {
 	NSRange		theRange = NSMakeRange( UINT_MAX, UINT_MAX );
-	unsigned int	theLenOfStart = [aStartString length],
+	NSUInteger	theLenOfStart = [aStartString length],
 						theLenOfEnd = [anEndString length];
 	
 	switch( aMode )
@@ -298,17 +298,17 @@
 	return theResult;
 }
 
-- (unsigned int)indexOfCharacater:(unichar)aChar
+- (NSUInteger)indexOfCharacater:(unichar)aChar
 {
 	return [self indexOfCharacater:(unichar)aChar options:0 range:NSMakeRange(0, [self length])];
 }
 
-- (unsigned int)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)anOptions
+- (NSUInteger)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)anOptions
 {
 	return [self indexOfCharacater:(unichar)aChar options:anOptions range:NSMakeRange(0, [self length])];
 }
 
-- (unsigned int)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)anOptions range:(NSRange)aRange
+- (NSUInteger)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)anOptions range:(NSRange)aRange
 {
 	return [self rangeOfString:[NSString stringWithCharacters:&aChar length:1] options:anOptions range:aRange].location;
 }
@@ -319,7 +319,7 @@
 - (NSArray *)componentsSeparatedByString:(NSString *)aSeparator withOpeningQuote:(NSString *)aOpeningQuote closingQuote:(NSString *)aClosingQuote singleQuote:(NSString *)aSingleQuote includeEmptyComponents:(BOOL)aFlag
 {
 	NSMutableArray		* theComponentArray = [NSMutableArray array];
-	unsigned int		theTokenEnd = 0,
+	NSUInteger          theTokenEnd = 0,
 						theLength = [self length],
 						theSeperatorLen = [aSeparator length],
 						theSingleQuoteLen = [aSingleQuote length],
@@ -399,7 +399,7 @@
 	NSCharacterSet		* theInverseSet = [aSet invertedSet];
 	NSRange				theStartRange = [self rangeOfCharacterFromSet:theInverseSet],
 						theEndRange = NSMakeRange( 1, 0 );
-	unsigned int		theLength = [self length];
+	NSUInteger          theLength = [self length];
 
 	while( theStartRange.location != NSNotFound )
 	{
